Firmware: Added a low-power release, suitable for chips with ckdiv8 programmed.
When it launches user program program, the chip is running at 8.0mhz, so can run at 3v within spec so long as PB3/D- isn't pulled high at startup (so skips boot loader) - so this works great if your usb circuitry is in a cable or dongle. If it accidentally runs the boot loader, it's only out of spec for 2.5 seconds.