Fix OSCCAL restor clash
A recent change broke the RESTORE_OSCCAL code. Fix that by disabling
the the OSCCAL restoration from the tiny table when RESTORE_OSCCAL is
defined (The osccal value will still be stored in the tiny table though).

Also added comment why the osccal value from the tiny table is not used
for the bootloader itself.