Get rid of the interrupt table, support for F_CPU != 16.5 MHz
Code size is now 2204 bytes, but that now includes the tiny table, so this
change effectively adds only 10 bytes.

In Detail:

Support for F_CPU != 16.5 MHz:
------------------------------

Micronucleus included a hacked vUSB version, that checks the stack content
to determine which ISR to call, that from the application or the one in the
bootloader. However, only the 16.5MHz and 12MHz versions had been adapted,
other frequencies did not work out of the box.

This change moves the stack inspection part into the bootloader, which allows
the use of any of the frequencies supported by vUSB (vUSB still needs a trivial
patch to export "waitForJ", normally an internal symbol).

In addition, there is now a compile-time switch (define APP_USES_VUSB) to tell
the bootloader that the application will use vUSB too. With that switch
enabled, the additional interrupt latency is reduced to a maximum 14 cycles,
which is well in the range allowed by vUSB. Without, the added interrupt
latency for the application is up to 23 cycles, which is pretty much on the
edge of what vUSB can tolerate.

Removed the interrupt table:
----------------------------

With some linker script fiddling, it is possible to get rid of the interrupt
table normally added to each program.

The Attiny 85 has 15 interrupts, which each consumes one word, so getting rid
of the table gains only 30 Bytes - not that much.

However, other devices from the ATTiny family have much larger interrupt
tables, the 167 has 20 interrupts and each vector needs 4 bytes.

Where the increased code size comes from:
-----------------------------------------

While getting rid of the interrupt table should spare some bytes, code still
got bigger. This is because it is now necessary to deal with function
addresses.

The actual storage address of a function is only known to the linker, thus all
address calculations will be deferred until linking. The linker knows about
several operations on addresses, for example how to construct a relative jump
or how to add a fixed offset when loading register , thus something like this
is not particularity expensive:

  void foo(void) {...}
  data = foo - 10 - 1;

which translates into something like

  ldi r24, lo8(gs(foo - 11))
  ldi r25, hi8(gs(foo - 11))

The second argument is inserted by the linker later.

This however does only work with +/-, other operations are not supported by
binutils for AVR (there is a feature called "Complex Relocation" for another
processor though), so a statement like

  data = (foo - 10 - 1) & 0xFFF) | C000;

will add least two more operations, which each costs a word.