Fix upgrade for 16k devices (?) and a reboot fix.
For some yet unknown reasons, initializing a global "bootloader_size" in main()
did not work properly for the '167. When accessing that variable from
"write_new_bootloader()", the size was only 48 bytes, although I found the
assembly code that calculated the size correctly.

Since the bootloader size is only used in "write_new_bootloader()", making it
a local variable there apparently fixed that.

The actual cause may have been somewhere else though. The "rebbot()" function
was defined as a little piece of assembly reading "rjmp 0" - witch is utterly
wrong, it says "jump one word forward", not "jump to the reset vector". Since
"reboot()" was not marked as "inline", the compiler moved to just before
"main()" and that might be what caused the problem in the first place. Not
entirely sure about that though.