Bring back USB disconnect in the upgrade firmware.
The previous code simply pulled down all the pins of port B for a while, then
released them again. The intention was probably to prevent USB bus
re-enumerations while upgrading the bootloader.

This would cause troubles when the device is still in its circuit and the pins
must not be manipulated. (in my case, some of the pins are used for I²C,
pulling down the pins will cause bus interruptions in this case).

The previous code also let the D- pin float after a while, which then caused a
bus re-enumeration before we entered the bootloader again.

This change
a) pulls down the D- pin only
b) does _not_ reset the pin configuration when leaving for the botloader.

The latter is so to prevent the host from enumerate the USB bus until the
bootloader is ready again.

D+ is left unchanged, it should be pulled down by the USB host anyways.