Get rid of tiny85FlashInit()
"tiny85FlashInit()" is only useful once, just after the bootloader has been
flashed via ISP, otherwise, the bootloader or "upgrade" would have inserted the
right vectors (reset, vUSB) already. If we just slap a vector table with the
correct vectors in front of .hex file, we do not need that function anymore.

With this, size for tiny-85 is 2112 bytes (-38).

Details:

"bootloader-vectors.S" provides the vector table. It is similar to what
"gcrt.o" from avr-libc does to produce the table, but it does not insert a call
to the vUSB interrupt (which is the one that is name __vector_X and would
normally be inserted into the vector table), but a call to our interrupt
wrapper. This vector table goes into a new section ".vectorpage".

The Makefile and the ld-scripts make the linker insert new ".vectorpage"
section at address 0 and also make it insert "rjmps" to the right spots in the
bootloader.

Finally, the Makefile rule to produce the .hex file now includes the new
section too, it ends up right at the start. The space between the new vector
table and the loader is left out, avrdude can handle that.

The construction of the upgrade firmware did not change. There we include a
raw dump of the bootloader, without the new vector table, just like before.