Fix usb disconnect bug and gain another 6 bytes.
When the upload utility sends the run command as a control message, it (resp.
the hosts USB stack) expects an ACK, but the ACK seems to get lost and the host
repots a broken pipe.

There was a 10ms delay loop in "leaveBootloader()" that appears to be there for
exactly this case, but it did not work for me, not even when I extended the
delay.

Calling "usbPoll()" again however seems to fix the missing ACK.

So this change

* removes the delay loop in "leaveBootloader()"
* adjusts "idlePolls" so, that two more calls to "usbPoll()" will happen.

This saves another 6 bytes.

Except, that this will not work in jumper mode. Here the bootloaderCondition is
always true and the break from the loop is absolutely necessary. Thus, if
BUILD_JUMPER_MODE is set, the old behavior - break the loop, then sleep 10ms - is
emulated by moving the delay into the main loop.

tiny-85 size now: 2102 bytes.