Put optional features into their own header files.
The features now export export their functionality as "static inline"
when enabled or as an empty define when not. This allows us to get
rid of many of the "#ifdef"s in function bodies, and makes them a bit
better to read.

This also includes the idle poll counting, which can now be disabled
completely, which save a few bytes in jumper mode.

The macros in "portnames.h" can be used to generate the special function
register names for port access. For example, PORT_DDR(B) expands simply
to DDRB, but B itself can be a macro too (LED_PORT ...)

Since now all the stuff that was formerly in bootloader.h is either somewhere
in hardware/ or in one of the new includes, bootloaderconfig.h has become
obsolete and has been removed.