Use (and fix) the LED support.
In the bootloader, the led is turned on when entering the main loop and turned
off when leaving. This coste 6 bytes for devices with a LED.

Upgrade will also turn on the LED when entering, but then toggles it for every
page written to flash. It also turns off the LED when leaving.

Added the LED defines for my '167 and '861 configurations.