libmnflash: allow redirecting library messages.
mnflash_set_message_log() and mnflash_set_error_log() can now be used
to intercept all messages in the library. Both take the address of a callback
function as argument which gets the error/output message for further processing
or printing.

This is useful if the library is integrated into a GUI application or a
background service.