bootloader: Add install target to the Makefile
"make install" installs the built hex and elf files on the system - by
default to "/usr/local/share/micronucleus/<target name>". The idea is
that the upload utility then can then find the right variant for a
given hardware.

The installation path can be influenced by passing "prefix" and "datarootdir"
as arguments to "make". "DESTDIR" is supported too.