Add debug logging facility
For completeness and because I need that elsewhere, it is not (yet) used by the
lib.

Debug logging is disabled by default but can be enabled through
"mnflash_set_debug_log()", which can be used to set a function that will do the
actual logging or by using "mnflash_enable_debug(1)", which makes the debug
output to use the same function as used for normal messages.