Fix two GCC 4.8 issues
a) move address increment in writeWordToPageBuffer() before resetting
  SREG. Without that, GCC 4.8 generates code to load "currentAddress"
  from RAM again.

b) make vUSB use switch/case instead of if/else
   With gcc <= 4.7, some switch statements caused more code to be produced then
   equivalent if/else statements and vUSB has special stuff to deal with that.
   With GCC 4.8, this seems no longer to be the case and using 'switch' is more
   efficient - at least in this case.

With this patch, GCC 4.8 can produce code that is a few bytes smaller then that
from 4.7, but only for devices with a crystal. Code for devices running on the
RC will gain another 20 or so bytes.