Introduce flag to set CKDIV=0 on bootloader entry
For certain applications there could be the need to operate the AVR on
low voltage, meaning it would not run properly on 8 MHz.
A flag CKDIV_RESET is introduced that runs after the bootloader entry
has been decided and clears CKDIV to 0.
Another flag CKDIV_RESTORE_DEFAULT allows resetting CKDIV to the reset value before entry of
the application, so the application sees the same clock speed as it does
without the bootloader being activated before.